requirements.txt
asyncpg==0.30.0
discord.py==2.4.0
hiredis==3.1.0
psutil==6.1.0
redis==5.2.1
tortoise-orm==0.22.1
